DAOS-5582 tse: zero the stack pointer in tse_task_reinit()
In common usage of task stack space, we push it in task func and pop
it in task's completion callback. But when user register its own
completion callback for that task and in side that reinit the task,
it will cause the other completion callback cannot be called and the
pushed stack space cannot be popped, then error happens when the
reinit-ed task re-executes.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>